Move uglify params to script definition
Follows-up 46ba111.

Instead of passing these as custom parameters to npm run,
it probably makes more sense to have them there by default
so that the command is always correct in the long term
even when the params or file names change again - without having
to check the readme each time.